Fixed warnings with MSVC compile.
FossilOrigin-Name: e0aa50c382a56bb02eaa589619b7e670006a26b3

+++ src/sqliteInt.h
@@ -1005,7 +1005,7 @@ struct FuncDef {
 */
 struct Savepoint {
   char *zName;                        /* Savepoint name (nul-terminated) */
-  int nDeferredCons;                  /* Number of deferred fk violations */
+  i64 nDeferredCons;                  /* Number of deferred fk violations */
   Savepoint *pNext;                   /* Parent savepoint (if any) */
 };
